prep for multi-return.

+++ source/m3_function.h
@@ -16,11 +16,11 @@ d_m3BeginExternC
 
 typedef struct M3FuncType
 {
-	struct M3FuncType *     next;
+    struct M3FuncType *     next;
 
-	u32                     numRets;
-	u32                     numArgs;
-	u8                      types [];        // returns, then args
+    u16                     numRets;
+    u16                     numArgs;
+    u8                      types [];        // returns, then args
 }
 M3FuncType;
 
@@ -29,7 +29,7 @@ typedef M3FuncType *        IM3FuncType;
 
 M3Result    AllocFuncType                   (IM3FuncType * o_functionType, u32 i_numTypes);
 bool        AreFuncTypesEqual               (const IM3FuncType i_typeA, const IM3FuncType i_typeB);
-u32			GetFuncTypeNumReturns			(const IM3FuncType i_funcType);
+u32         GetFuncTypeNumReturns           (const IM3FuncType i_funcType);
 
 
 //---------------------------------------------------------------------------------------------------------------------------------
@@ -82,7 +82,7 @@ void        Function_FreeCompiledCode   (IM3Function i_function);
 cstr_t      GetFunctionImportModuleName (IM3Function i_function);
 cstr_t *    GetFunctionNames            (IM3Function i_function, u16 * o_numNames);
 u32         GetFunctionNumArgs          (IM3Function i_function);
-u8          GetFunctionArgType   	    (IM3Function i_function, u32 i_index);
+u8          GetFunctionArgType          (IM3Function i_function, u32 i_index);
 
 u32         GetFunctionNumReturns       (IM3Function i_function);
 u8          GetFunctionReturnType       (IM3Function i_function, u32 i_index);
